changes for vs2017

+++ src/ctlinedata2georefimage.cpp
@@ -815,8 +815,8 @@ int main(int argc, char** argv)
 
 	std::vector<std::string> filelist =  cDirectoryAccess::getfilelist(infiles);
 	double t1 = gettime();	
-	for (int i = 0; i < filelist.size(); i++){	
-		printf("Processing file %s %3d of %3d\n", filelist[i].c_str(),i+1,filelist.size());
+	for (size_t i = 0; i < filelist.size(); i++){	
+		printf("Processing file %s %3zu of %3zu\n", filelist[i].c_str(),i+1,filelist.size());
 		cGeorefSection S;		
 		S.getoptions(sectionblock);		
 		S.readdatafile(inputblock,filelist[i].c_str());
